Move waitForOpenCV to its own util; Rename utils to tool-utils
