+++ web_ui/packages/smart-tools/src/utils/tool-utils.ts
