+++ web_ui/src/webworkers/inference-image.worker.ts
@@ -1,30 +1,11 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { InferenceImage, OpenCVLoader } from '@geti/smart-tools';
-import { expose, proxy } from 'comlink';
-import type OpenCVTypes from 'OpenCVTypes';
+import { InferenceImage, waitForOpenCV } from '@geti/smart-tools';
+import { expose } from 'comlink';
 
 declare const self: DedicatedWorkerGlobalScope;
 
-let opencv: OpenCVTypes;
-
-const waitForOpenCV = async () => {
-    if (opencv) return true;
-
-    opencv = await OpenCVLoader();
-
-    if ('ready' in opencv) {
-        await opencv.ready;
-    }
-
-    return false;
-};
-
-const loadInferenceImage = async () => {
-    return proxy(new InferenceImage(opencv));
-};
-
-const WorkerApi = { InferenceImage: loadInferenceImage, waitForOpenCV, terminate: self.close };
+const WorkerApi = { InferenceImage, waitForOpenCV, terminate: self.close };
 
 expose(WorkerApi);
